feat: Basic Setup of Home Assistant
I need a control center for my smart home. For this I would like to use Homeassistant, which can integrate many useful interfaces.
This commit contains a first simple deployment of HAS and a simple config.

A lot of things are configured in Homeassistant UI and some can only be set via configuration.yaml, so added to git.
For later extensions I need an MQTT broker, so it is already part of this version.