tools: Backup script for the home central
Of course I would like to back up data from this control center. For this reason I added a script.
The script backs up all data that is created at runtime of the services and backs it up to another path of the local file system.
This location is backed up by another mechanic.

Only the data that is created at runtime is backed up, configurations should be backed up by this git.