feat: Read out the timer of a shelly plug via rest
I have a shelly plug that is automatically switched off after a timer
expires. Unfortunate the remaining is not provided by the shelly plugin
of HAS. This Value has to collect via rest.

This Commit adds a new rest call to the configuration for getting this
value. The Value is send in s and I convert this to minutes.