feat: Control devices via zigbee
I have several devices that run on Zigbee and I want to avoid having
a gateway from Phillips or Ikea because it needs a place to mount on it and some
features are not supported by them.
Luckily for me I found this nice tool zigbee2mqtt. Since I have
connected a Zigbee Stick to my server I can control all devices with
zigbee and connect them via MQTT to HAS.

I just pick the official docker compose entry from zigbee2mqtt.
By this commit I just change the device of the Container to the ID of my
devices and I change the restart policy to always to be consistent with
the other containers.