fix: Missing mqtt config with listener
In general you need to configure a listener in mosquitto,if you want to
open the MQTT Port externally. This configuration was already done by
me. During the test of last commits, if have seen that this file is not
added. By this commit I added the config file for mosquitto.